+++ Backend/src/main/java/pl/smarthome/Repositories/ComponentRepository.java
@@ -6,6 +6,6 @@ import pl.smarthome.Models.Component;
 import java.util.Optional;
 
 public interface ComponentRepository  extends JpaRepository<Component, Long> {
-    Optional<Component> findById(Integer id);
+    Optional<Component> findById(Long id);
     Optional<Component> findByName(String name);
 }
